Added combine reducers for splitting UI and machine actions
