+++ src/interface/components/Superescalar/modal/AutorModalComponent.tsx
@@ -47,7 +47,7 @@ class AutorModalComponent extends React.Component<any, any> {
 
 const mapStateToProps = state => {
       return {
-          isAuthorModalOpen: state.isAuthorModalOpen,
+          isAuthorModalOpen: state.Ui.isAuthorModalOpen,
       }
   }
