+++ src/interface/components/Superescalar/modal/LoadModalComponent.tsx
@@ -96,7 +96,7 @@ LOOP:
 
 const mapStateToProps = state => {
     return {
-        isLoadModalOpen: state.isLoadModalOpen,
+        isLoadModalOpen: state.Ui.isLoadModalOpen,
     }
 }
