+++ src/interface/components/Superescalar/modal/OptionsModalComponent.tsx
@@ -71,7 +71,7 @@ class OptionsModalComponent extends React.Component<any, any> {
 
 const mapStateToProps = state => {
     return {
-        isOptionsModalOpen: state.isOptionsModalOpen,
+        isOptionsModalOpen: state.Ui.isOptionsModalOpen,
     }
 }
