+++ src/interface/components/VLIW/modal/VLIWLoadModalComponent.tsx
@@ -135,7 +135,7 @@ LOOP:
 
 const mapStateToProps = state => {
     return {
-        isLoadModalOpen: state.isLoadModalOpen,
+        isLoadModalOpen: state.Ui.isLoadModalOpen,
     }
 }
